Refactor hooks and utils
