+++ apps/tup-ui/src/hooks/index.ts
@@ -1,0 +1,56 @@
+
+import { setLogger } from 'react-query';
+
+// Disable error logging when we throw inside a react-query fetcher method.
+setLogger({
+  log: window.console.log,
+  warn: window.console.warn,
+  error: () => {
+    /* */
+  },
+});
+
+// Extend the window type to support custom config passed from the server side.
+declare global {
+  interface Window {
+    __TUP_CONFIG__: {
+      baseUrl: string;
+    };
+  }
+}
+
+export type UserProfile = {
+  id: number;
+  username: string;
+  email: string;
+  firstName: string;
+  lastName: string;
+  institution: string;
+  institutionId: number;
+  department: string;
+  departmentId: number;
+  country: string;
+  countryId: number;
+  citizenship: string;
+  citizenshipId: number;
+  piEligibility: string;
+  phone: string;
+  title: string;
+  uid: number;
+  homeDirectory: string;
+  gid: number;
+  emailConfirmations: string[];
+};
+
+export type AuthResponse = {
+  exp: number;
+  iat: number;
+  ttl: number;
+  jwt: string;
+  username: string;
+  profile: UserProfile;
+};
+
+export { default as useAuthMutation } from './useAuthMutation';
+export { default as useProfileQuery } from './useProfileQuery';
+export { default as useJwt } from './useJwt';
